rename: repository 테스트 패키지 이름 변경
