+++ src/test/java/roomescape/repository/ReservationRepositoryTest.java
@@ -1,4 +1,4 @@
-package roomescape.dao;
+package roomescape.repository;
 
 import static org.assertj.core.api.Assertions.assertThat;
